Fixed Issues with Rev 1 Schematic and PCB
Connections between L293S inputs and RPi where grounded (physical pins 30 and 43). These where moved to 29 and 31 on the opposite side of the GPIO header. Also includes the updated PCB design files and ArtWorks. Also fixed a track which connected to the PCB header on the bottom side (which was awkward to solder), moved this track to the top side of the etching layer.